+++ apps/desktop/src/components/helpers/dev-tools.tsx
@@ -1,7 +1,6 @@
 import { usePersistedStore } from '@/lib/store';
 import { invoke } from '@tauri-apps/api/core';
 import { Button } from '../ui/button';
-import TailwindIndicator from './tailwind-indicator';
 
 const DevTools = () => {
   const { clearMods } = usePersistedStore();
@@ -10,7 +9,6 @@ const DevTools = () => {
     <div className="flex gap-2">
       <Button onClick={clearMods}>Clear Mods</Button>
       <Button onClick={() => invoke('stop_game')}>Stop Game</Button>
-      <TailwindIndicator />
     </div>
   );
 };