+++ apps/desktop/src/components/menu.tsx
@@ -12,6 +12,7 @@ import {
 } from '@/components/ui/menubar';
 
 import { getCurrentWindow } from '@tauri-apps/api/window';
+import { open } from '@tauri-apps/plugin-shell';
 import { AboutDialog } from './about-dialog';
 import Logo from './logo';
 import { Dialog, DialogTrigger } from './ui/dialog';
@@ -57,6 +58,15 @@ export function Menu() {
             </DialogTrigger>
             <AboutDialog />
           </Dialog>
+        </MenubarMenu>
+
+        <MenubarMenu>
+          <MenubarTrigger className="relative text-sm">Help</MenubarTrigger>
+          <MenubarContent>
+            <MenubarItem onClick={() => open('https://discord.gg/KSB2kzQWWE')}>
+              Need Help? Join Discord
+            </MenubarItem>
+          </MenubarContent>
         </MenubarMenu>
       </Menubar>
     </WindowTitlebar>